requirements.txt
pySerial
tqdm